restructuring / tidy up
