+++ DuckDuckGo/FavoritesHomeViewSectionRenderer.swift
@@ -1,0 +1,66 @@
+//
+//  FavoritesHomeViewSectionRenderer.swift
+//  DuckDuckGo
+//
+//  Created by Chris Brind on 26/11/2018.
+//  Copyright © 2018 DuckDuckGo. All rights reserved.
+//
+
+import UIKit
+
+class FavoritesHomeViewSectionRenderer: HomeViewSectionRenderer {
+    
+    // Delegate to model, plus one for the plus button
+    private let numberOfItems = 9
+    
+    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
+        return numberOfItems
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+        
+        if isLastItem(indexPath) {
+            return collectionView.dequeueReusableCell(withReuseIdentifier: "addFavorite", for: indexPath)
+        } else {
+            guard let view = collectionView.dequeueReusableCell(withReuseIdentifier: "favorite", for: indexPath) as? FavoriteHomeCell else {
+                fatalError("not a FavoriteCell")
+            }
+            return view
+        }
+        
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, layout collectionViewLayout: UICollectionViewLayout, sizeForItemAt indexPath: IndexPath)
+        -> CGSize {
+            return CGSize(width: 80, height: 100)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, canMoveItemAt indexPath: IndexPath) -> Bool {
+        print("***", #function, indexPath)
+        return !isLastItem(indexPath)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, moveItemAt sourceIndexPath: IndexPath, to destinationIndexPath: IndexPath) {
+        print("***", #function)
+    }
+    
+    func collectionView(_ collectionView: UICollectionView, targetIndexPathForMoveFromItemAt originalIndexPath: IndexPath,
+                        toProposedIndexPath proposedIndexPath: IndexPath) -> IndexPath {
+        print("***", #function, originalIndexPath, proposedIndexPath)
+        guard originalIndexPath.section == proposedIndexPath.section else { return originalIndexPath }
+        guard !isLastItem(proposedIndexPath) else { return originalIndexPath }
+        return proposedIndexPath
+    }
+    
+    func menuItemsFor(itemAt: Int) -> [UIMenuItem] {
+        return [
+            UIMenuItem(title: UserText.favoriteMenuDelete, action: FavoriteHomeCell.Actions.delete),
+            UIMenuItem(title: UserText.favoriteMenuEdit, action: FavoriteHomeCell.Actions.edit)
+        ]
+    }
+    
+    private func isLastItem(_ indexPath: IndexPath) -> Bool {
+        return indexPath.row + 1 == numberOfItems
+    }
+    
+}
